# Conflicts:
#	scripts/docker-entrypoint.sh
#	scripts/env-data.sh

+++ scripts/env-data.sh
@@ -111,8 +111,8 @@ DATA_PATH=$1
 
 if [[ ! -d ${DATA_PATH} ]];
 then
-    echo "Creating" ${DATA_PATH}  "directory"
-    mkdir -p ${DATA_PATH}
+    echo "Creating" "${DATA_PATH}"  "directory"
+    mkdir -p "${DATA_PATH}"
 fi
 }
 
@@ -367,11 +367,21 @@ if [ -z "$EXTRA_CONF" ]; then
     EXTRA_CONF=""
 fi
 
+if [ -z "$ACTIVATE_CRON" ]; then
+    ACTIVATE_CRON=TRUE
+fi
+
 if [ -z "${SHARED_PRELOAD_LIBRARIES}" ]; then
     if [[ $(dpkg -l | grep "timescaledb") > /dev/null ]];then
-        SHARED_PRELOAD_LIBRARIES='pg_cron,timescaledb'
+        if [[ ${ACTIVATE_CRON} =~ [Tt][Rr][Uu][Ee] ]];then
+          SHARED_PRELOAD_LIBRARIES='pg_cron,timescaledb'
+        else
+          SHARED_PRELOAD_LIBRARIES='timescaledb'
+        fi
     else
-        SHARED_PRELOAD_LIBRARIES='pg_cron'
+        if [[ ${ACTIVATE_CRON} =~ [Tt][Rr][Uu][Ee] ]];then
+          SHARED_PRELOAD_LIBRARIES='pg_cron'
+        fi
     fi
 fi
 
@@ -420,9 +430,10 @@ if [ -n "${POSTGRES_INITDB_ARGS}" ]; then
   INITDB_EXTRA_ARGS=${POSTGRES_INITDB_ARGS}
 fi
 
-list=(`echo ${POSTGRES_DBNAME} | tr ',' ' '`)
-arr=(${list})
-SINGLE_DB=${arr[0]}
+IFS=','
+read -a dbarr <<< "$POSTGRES_DBNAME"
+SINGLE_DB=${dbarr[0]}
+export ${SINGLE_DB}
 
 if [ -z "${TIMEZONE}" ]; then
   TIMEZONE='Etc/UTC'
@@ -430,12 +441,12 @@ fi
 
 # usable function definitions
 function kill_postgres {
-  PID=`cat ${PG_PID}`
-  kill -TERM ${PID}
+  PID=$(cat "${PG_PID}")
+  kill -TERM "${PID}"
 
   # Wait for background postgres main process to exit
   # wait until PID file gets deleted
-  while ls -A ${PG_PID} 2> /dev/null; do
+  while ls -A "${PG_PID}" 2> /dev/null; do
     sleep 1
   done
 
@@ -465,6 +476,8 @@ function restart_postgres {
 
 function entry_point_script {
   SETUP_LOCKFILE="${SCRIPTS_LOCKFILE_DIR}/.entry_point.lock"
+  IFS=','
+  read -a dbarr <<< "$POSTGRES_DBNAME"
   # If lockfile doesn't exists, proceed.
   if [[ ! -f "${SETUP_LOCKFILE}" ]] || [[ "${IGNORE_INIT_HOOK_LOCKFILE}" =~ [Tt][Rr][Uu][Ee] ]]; then
       if find "/docker-entrypoint-initdb.d" -mindepth 1 -print -quit 2>/dev/null | grep -q .; then
@@ -473,21 +486,21 @@ function entry_point_script {
           case "$f" in
                 *.sql)    echo "$0: running $f";
                   if [[ "${ALL_DATABASES}" =~ [Ff][Aa][Ll][Ss][Ee] ]]; then
-                      psql ${SINGLE_DB} -U ${POSTGRES_USER} -p 5432 -h localhost  -f ${f} || true
+                      psql "${SINGLE_DB}" -U ${POSTGRES_USER} -p 5432 -h localhost  -f "${f}" || true
                   else
-                      for db in $(echo ${POSTGRES_DBNAME} | tr ',' ' '); do
-                        psql ${db} -U ${POSTGRES_USER} -p 5432 -h localhost  -f ${f} || true
+                      for db in "${dbarr[@]}";do
+                        psql "${db}" -U ${POSTGRES_USER} -p 5432 -h localhost  -f "${f}" || true
                       done
                   fi;;
                 *.sql.gz) echo "$0: running $f";
                   if [[ "${ALL_DATABASES}" =~ [Ff][Aa][Ll][Ss][Ee] ]]; then
-                      gunzip < "$f" | psql ${SINGLE_DB} -U ${POSTGRES_USER} -p 5432 -h localhost || true
+                      gunzip < "$f" | psql "${SINGLE_DB}" -U ${POSTGRES_USER} -p 5432 -h localhost || true
                   else
-                      for db in $(echo ${POSTGRES_DBNAME} | tr ',' ' '); do
-                        gunzip < "$f" | psql ${db} -U ${POSTGRES_USER} -p 5432 -h localhost || true
+                      for db in "${dbarr[@]}";do
+                        gunzip < "$f" | psql "${db}" -U ${POSTGRES_USER} -p 5432 -h localhost || true
                       done
                   fi;;
-                *.sh)     echo "$0: running $f"; . $f || true;;
+                *.sh)     echo "$0: running $f"; . "$f" || true;;
                 *)        echo "$0: ignoring $f" ;;
             esac
             echo
@@ -511,8 +524,8 @@ function configure_replication_permissions {
       non_root_permission "${USER_NAME}" "${DB_GROUP_NAME}"
 
     else
-      chown -R postgres:postgres ${DATADIR} ${WAL_ARCHIVE}
-      chmod -R 750 ${DATADIR} ${WAL_ARCHIVE}
+      chown -R postgres:postgres "${DATADIR}" ${WAL_ARCHIVE}
+      chmod -R 750 "${DATADIR}" ${WAL_ARCHIVE}
       echo -e "[Entrypoint] \e[1;31m Setup data permissions for replication as root user \033[0m"
       chown -R postgres:postgres $(getent passwd postgres | cut -d: -f6)
       su - postgres -c "echo \"${REPLICATE_FROM}:${REPLICATE_PORT}:*:${REPLICATION_USER}:${REPLICATION_PASS}\" > ~/.pgpass"
@@ -521,13 +534,13 @@ function configure_replication_permissions {
 }
 
 function streaming_replication {
-  until ${START_COMMAND} "${PG_BASEBACKUP} -X stream -h ${REPLICATE_FROM} -p ${REPLICATE_PORT} -D ${DATADIR} -U ${REPLICATION_USER} -R -vP -w --label=gis_pg_custer"
+  until START_COMMAND "${PG_BASEBACKUP} -X stream -h ${REPLICATE_FROM} -p ${REPLICATE_PORT} -D ${DATADIR} -U ${REPLICATION_USER} -R -vP -w --label=gis_pg_custer"
     do
       echo -e "[Entrypoint] \e[1;31m Waiting for master to connect... \033[0m"
       sleep 1s
-      if [[ "$(ls -A ${DATADIR})" ]]; then
+      if [[ "$(ls -A "${DATADIR}")" ]]; then
         echo -e "[Entrypoint] \e[1;31m Need empty folder. Cleaning directory... \033[0m"
-        rm -rf ${DATADIR}/*
+        rm -rf "${DATADIR:?}/"*
       fi
     done
 
@@ -557,20 +570,21 @@ function over_write_conf() {
 
 function extension_install() {
   DATABASE=$1
+  DB_EXTENSION=$2
   IFS=':'
-  read -a strarr <<< "$ext"
+  read -a strarr <<< "${DB_EXTENSION}"
   EXTENSION_NAME=${strarr[0]}
   EXTENSION_VERSION=${strarr[1]}
   if [[ -z ${EXTENSION_VERSION} ]];then
     if [[ ${EXTENSION_NAME} != 'pg_cron' ]]; then
       echo -e "\e[32m [Entrypoint] Enabling extension \e[1;31m ${EXTENSION_NAME} \e[32m in the database : \e[1;31m ${DATABASE} \033[0m"
-      psql ${DATABASE} -U ${POSTGRES_USER} -p 5432 -h localhost -c "CREATE EXTENSION IF NOT EXISTS \"${EXTENSION_NAME}\" cascade;"
+      psql "${DATABASE}" -U ${POSTGRES_USER} -p 5432 -h localhost -c "CREATE EXTENSION IF NOT EXISTS \"${EXTENSION_NAME}\" cascade;"
     fi
   else
     if [[ ${EXTENSION_NAME} != 'pg_cron' ]]; then
       pattern="${EXTENSION_NAME}--"
       last_numbers=()
-      for file in "$EXTDIR"/${pattern}*; do
+      for file in "$EXTDIR"/"${pattern}"*; do
         filename=$(basename "$file" .sql)
         if [[ "$filename" == *"--"* ]]; then
           last_number=$(echo "$filename" | awk -F '--' '{print $NF}')
@@ -581,7 +595,7 @@ function extension_install() {
       done
       if [[ " ${last_numbers[@]} " =~ " $EXTENSION_VERSION " ]]; then
         echo -e "\e[32m [Entrypoint] Installing extension \e[1;31m ${EXTENSION_NAME}  \e[32m with version \e[1;31m ${EXTENSION_VERSION} \e[32m in the database : \e[1;31m ${DATABASE} \033[0m"
-        psql ${DATABASE} -U ${POSTGRES_USER} -p 5432 -h localhost -c "CREATE EXTENSION IF NOT EXISTS \"${EXTENSION_NAME}\" WITH VERSION '${EXTENSION_VERSION}' cascade;"
+        psql "${DATABASE}" -U ${POSTGRES_USER} -p 5432 -h localhost -c "CREATE EXTENSION IF NOT EXISTS \"${EXTENSION_NAME}\" WITH VERSION '${EXTENSION_VERSION}' cascade;"
       else
         echo -e "\e[32m [Entrypoint] Extension \e[1;31m ${EXTENSION_NAME}  \e[32m with version \e[1;31m ${EXTENSION_VERSION} \e[32m is not available for install, available versions to install are \e[1;31m  "${last_numbers[@]}" \033[0m"
       fi
@@ -593,11 +607,11 @@ function extension_install() {
 
 function directory_checker() {
   DATA_PATH=$1
-  if [ -d $DATA_PATH ];then
-    DB_USER_PERM=$(stat -c '%U' ${DATA_PATH})
-    DB_GRP_PERM=$(stat -c '%G' ${DATA_PATH})
+  if [ -d "$DATA_PATH" ];then
+    DB_USER_PERM=$(stat -c '%U' "${DATA_PATH}")
+    DB_GRP_PERM=$(stat -c '%G' "${DATA_PATH}")
     if [[ ${DB_USER_PERM} != "${USER}" ]] &&  [[ ${DB_GRP_PERM} != "${GROUP}"  ]];then
-      chown -R ${USER}:${GROUP} ${DATA_PATH}
+      chown -R "${USER}":"${GROUP}" "${DATA_PATH}"
     fi
   fi
 
@@ -613,9 +627,9 @@ function non_root_permission() {
   done
   services=("/usr/lib/postgresql/" "/etc/" "/var/run/!(secrets)" "/var/lib/" "/usr/bin" "/tmp" "/scripts")
   for paths in "${services[@]}"; do
-    directory_checker $paths
+    directory_checker "${paths}"
   done
-  chmod -R 750 ${DATADIR} ${WAL_ARCHIVE}
+  chmod -R 750 "${DATADIR}" ${WAL_ARCHIVE}
 
 }
